Bvh stuff

+++ Win3D/src/scene/objects/PointLight.hpp
@@ -5,5 +5,5 @@
 struct PointLight {
     Vector position;
     Colour colour;
-    float intensity;
+    double intensity;
 };
